✨ Update New Page
