--- apps/web/src/app/[locale]/(admin)/new/page.tsx
@@ -1,46 +1,0 @@
-import { Button } from "@rallly/ui/button";
-import { Icon } from "@rallly/ui/icon";
-import { ArrowLeftIcon } from "lucide-react";
-import Link from "next/link";
-import { Trans } from "react-i18next/TransWithoutContext";
-
-import { Params } from "@/app/[locale]/types";
-import {
-  PageContainer,
-  PageContent,
-  PageHeader,
-} from "@/app/components/page-layout";
-import { getTranslation } from "@/app/i18n";
-import { CreatePoll } from "@/components/create-poll";
-
-export default async function Page({ params }: { params: Params }) {
-  const { t } = await getTranslation(params.locale);
-  return (
-    <PageContainer>
-      <PageHeader>
-        <Button asChild>
-          <Link href="/polls">
-            <Icon>
-              <ArrowLeftIcon />
-            </Icon>
-            <Trans i18nKey="back" t={t} defaults="Back" />
-          </Link>
-        </Button>
-      </PageHeader>
-      <PageContent>
-        <CreatePoll />
-      </PageContent>
-    </PageContainer>
-  );
-}
-
-export async function generateMetadata({
-  params,
-}: {
-  params: { locale: string };
-}) {
-  const { t } = await getTranslation(params.locale);
-  return {
-    title: t("newPoll"),
-  };
-}
