+++ apps/web/src/app/[locale]/new/loading.tsx
